Fixed the typo error and moved the auth logic in the auth_provider
